+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - Rename `deep` prop to `depth`
 - Make callback functions pass an object as the argument rather than multiple positional arguments
 - Rename `onDeltaCallback`'s `type` property to `dataType`
-- Make the `dataType` property in callback functions an enum `ObjectType`
+- Create new enums: 
+    - `dataType` property in callback functions is now `ObjectType`
+    - `inputUsageType` property in callback functions is now `JsonFieldType`
 
 ## Code
 - Update dependencies